% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyDiffs.R
\name{applyDiffs}
\alias{applyDiffs}
\title{Aplica as diferenciações necessárias à série temporal}
\usage{
applyDiffs(serie_temporal, d, D)
}
\description{
Aplica as diferenciações necessárias à série temporal
}
