% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineDiffs.R
\name{determineDiffs}
\alias{determineDiffs}
\title{determineDiffs}
\usage{
determineDiffs(serie_temporal)
}
\value{
Retorna o numero de diferença para retirar a tendencia e a sazonalidade
}
\description{
determineDiffs
}
