% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operador.R
\name{null_coalesce}
\alias{null_coalesce}
\alias{\%||\%}
\title{Operador de coalescência para NULL}
\usage{
a \%||\% b
}
\arguments{
\item{a}{Objeto a ser testado.}

\item{b}{Valor alternativo se \code{a} for \code{NULL}.}
}
\value{
\code{a} se não for \code{NULL}; caso contrário \code{b}.
}
\description{
Retorna \code{a} se não for \code{NULL}; caso contrário retorna \code{b}.
}
